antman: Remove `--checkupdate` flag
This was a redundant flag that pretty much did the same task as --update flag but had a seperate function for itself.
Remove this and rather add a confirmation input to the update function to provide the ability to cancel the update after check.

This commit only removes the checkupdate flag, above mentioned adjustments to --update flag will be coming in future commits.

Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>